2002-02-10  Daniel Jacobowitz  <drow@mvista.com>
        * coff-rs6000.c (xcoff_generate_rtinit): Silence uninitialized
        variable warnings.
        * elf32-sh.c (sh_elf_relax_section): Silence signed/unsigned
        comparison warning.
        * trad-core.c (trad_unix_core_file_p): Silence pointer/integer
        cast warnings for the common case.

+++ bfd/elf32-sh.c
@@ -2066,7 +2066,8 @@ sh_elf_relax_section (abfd, sec, link_info, again)
 	  shndx = shndx_buf + (shndx_buf ? ELF32_R_SYM (irelfn->r_info) : 0);
 	  bfd_elf32_swap_symbol_in (abfd, esym, shndx, &isym);
 
-	  if (isym.st_shndx != _bfd_elf_section_from_bfd_section (abfd, sec))
+	  if (isym.st_shndx
+	      != (unsigned int) _bfd_elf_section_from_bfd_section (abfd, sec))
 	    {
 	      ((*_bfd_error_handler)
 	       (_("%s: 0x%lx: warning: symbol in unexpected section"),
